+++ NEWS.md
@@ -1,7 +1,9 @@
 
 # gh development version
 
-* Token validation accounts for the new format [announced 2021-03-04 ](https://github.blog/changelog/2021-03-04-authentication-token-format-updates/) and implemented on 2021-04-01 (#148, @fmichonneau).
+* Token validation accounts for the new format
+  [announced 2021-03-04 ](https://github.blog/changelog/2021-03-04-authentication-token-format-updates/)
+  and implemented on 2021-04-01 (#148, @fmichonneau).
 
 # gh 1.2.0