+++ NEWS.md
@@ -1,3 +1,5 @@
+# gh 1.2.1
+
 
 # gh development version